Fixed the build errors (on Linux): * Wrong format string types. * Default argument values should be in header, not .cpp.

+++ Source/Readers/ImageReader/ImageDataDeserializer.cpp
@@ -73,7 +73,7 @@ ImageDataDeserializer::ImageDataDeserializer(const ConfigParameters& config)
     }
     else
     {
-        RuntimeError("Unsupported label element type '%d'.", label->m_elementType);
+        RuntimeError("Unsupported label element type '%d'.", (int)label->m_elementType);
     }
 
     CreateSequenceDescriptions(configHelper.GetMapPath(), labelDimension);
